BayesHMM: Full Bayesian Inference for Hidden Markov Models

An R Package to run full Bayesian inference on Hidden Markov Models (HMM) using the probabilistic programming language Stan. The software enables users to fit HMM with time-homogeneous transitions as well as time-varying transition probabilities. Priors can be set for every model parameter. Implemented inference algorithms include forward (filtering), forward-backwards (smoothing), Viterbi (most likely hidden path), prior predictive sampling, and posterior predictive sampling. Graphs, tables and other convenience methods for convergence diagnosis, goodness of fit, and data analysis are provided.

Details

See the Introduction vignette: vignette("introduction", package = "BayesHMM") Additionally, you may start with the manual help for ?specify and ?fit.

See also