Bernoulli mass (univariate, discrete, bounded space)

It can be used to specify either a prior distribution for a model parameter or a likelihood function for an observation model.

Bernoulli(theta = NULL, bounds = list(NULL, NULL), trunc = list(NULL,
  NULL), k = NULL, r = NULL, param = NULL)

Arguments

theta
Either a fixed value or a prior density for the success proportion parameter.

bounds
(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.

trunc
(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.

k
(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.

r
(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.

param
(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.

Value

A Density object.

See also

Examples

# With fixed values for the parameters
Bernoulli(0.5)

#> $name
#> [1] "Bernoulli"
#> 
#> $theta
#> [1] 0.5
#> 
#> $bounds
#> $bounds[[1]]
#> NULL
#> 
#> $bounds[[2]]
#> NULL
#> 
#> 
#> $trunc
#> $trunc[[1]]
#> NULL
#> 
#> $trunc[[2]]
#> NULL
#> 
#> 
#> $k
#> NULL
#> 
#> $r
#> NULL
#> 
#> $param
#> NULL
#> 
#> attr(,"class")
#> [1] "Bernoulli"       "DiscreteDensity" "Density"        

# With priors for the parameters
Bernoulli(
  Beta(1, 1)
)

#> $name
#> [1] "Bernoulli"
#> 
#> $theta
#> $name
#> [1] "Beta"
#> 
#> $alpha
#> [1] 1
#> 
#> $beta
#> [1] 1
#> 
#> $bounds
#> $bounds[[1]]
#> NULL
#> 
#> $bounds[[2]]
#> NULL
#> 
#> 
#> $trunc
#> $trunc[[1]]
#> NULL
#> 
#> $trunc[[2]]
#> NULL
#> 
#> 
#> $k
#> NULL
#> 
#> $r
#> NULL
#> 
#> $param
#> NULL
#> 
#> attr(,"class")
#> [1] "Beta"    "Density"
#> 
#> $bounds
#> $bounds[[1]]
#> NULL
#> 
#> $bounds[[2]]
#> NULL
#> 
#> 
#> $trunc
#> $trunc[[1]]
#> NULL
#> 
#> $trunc[[2]]
#> NULL
#> 
#> 
#> $k
#> NULL
#> 
#> $r
#> NULL
#> 
#> $param
#> NULL
#> 
#> attr(,"class")
#> [1] "Bernoulli"       "DiscreteDensity" "Density"