Extract the simulated sample of the observation variable (ysim).

If the fit object was created via the sim method, this function returns a sample of the observation variable from the prior predictive densitiy (i.e. using random values of the parameters). If a dataset was provided when fitting the model instead, the sample is simulated from the posterior predictive density. In the latter case, calling either extract_ysim and extract_ypred are equivalent.

extract_ysim(fit, n = NULL, ...)

Arguments

fit
An object returned by either sampling or optimizing.

Value

A numeric array with four dimensions [N, M, T, R]: number of iterations N, number of chains M, length of the time series T, the dimension of the observation vector R. If the argument chain was set, the chain dimension is dropped and the function returns a three-dimensional numeric array [N, T, R]. If the argument reduce was set to a function returning a vector of size n, the returned array has dimension [n, M, T, R]. When n is simply one (e.g. median), the number-of-iterations dimension is dropped and the function returns a three-dimensional numeric array [M, T, R].

See also