Draw samples from a specification.

Draw samples from a specification.

function(spec, stanModel = NULL, y, x = NULL, u = NULL, v = NULL, writeDir = tempdir(), ...)

Arguments

spec
An object returned by either specify or hmm.

...
Arguments to be passed to rstan's sampling.

stanModel
An optional instance of S4 class stanmodel returned by compile. If not given, the model is automatically compiled but the object is not returned to the user and cannot be reutilized in future sampling.

y
A numeric matrix with the observation sample. It must have as many rows as the time series length T and as many columns as the dimension of the observation vector R. If not a matrix, the function tries to cast the object to a \(T\times R\) matrix.

x
An optional numeric matrix with the covariates for the observation model. It must have as many rows as the time series length T and as many columns as the dimension of the covariate vector M. If not a matrix, the function tries to cast the object to a \(T\times M\) matrix. Useful for Hidden Markov Regression Model (also known as Markov-switching regressions).

u
An optional numeric matrix with the covariates for the transition model. It must have as many rows as the time series length T and as many columns as the dimension of the transition covariate vector P. If not a matrix, the function tries to cast the object to a \(T\times P\) matrix. Useful for Hidden Markov Models with time-varying transition probabilities.

v
An optional numeric matrix with the covariates for the initial distribution model. It must have as many rows as the number of hidden states K and as many columns as the dimension of the initial covariate vector Q. If not a matrix, the function tries to cast the object to a \(K\times Q\) matrix.

writeDir
An optional string with the path where the Stan file should be written. Useful to inspect and modify the Stan code manually. It defaults to a temporary directory.

Value

An object of S4 class stanfit with some additional attributes (the dataset data, the name of the Stan code file filename, and the specification object spec). This object is completely compatible with all other functions.

See also