Theme for BayesHMM visualizations and printouts

BayesHMM comes with many built-in visualization and text printout methods to carry out bayesian data analysis out of the box. All plot colors and printout text formats are based on a fully-customizable theme loaded as a global setting. The theme is simply a named list with two elements BayesHMM.theme and BayesHMM.print, which in turn are named lists following the structure described below. Run get_default_theme to see the default values.

Details

BayesHMM.theme: settings for visualizations.

boxY

A string with the color of the lines of the box corresponding to the observed sample. Applies to plot_ppredictive with features boxplot.

boxYPred

A string with the color of the lines of the many boxes corresponding to the samples drawn from the posterior predictive density. Applies to plot_ppredictive with features boxplot.

cumulativeY

A string with the color of the cumulative density line for the observed sample. Applies to plot_ppredictive with features cumulative.

cumulativeYPred

A string with the color of the cumulative density lines for the many samples drawn from the posterior predictive density. Applies to plot_ppredictive with features cumulative.

densityY

A string with the color of the density line for the observed sample. Applies to plot_ppredictive with features density.

densityYPred

A string with the color of the density lines for the many samples drawn from the posterior predictive density. Applies to plot_ppredictive with features density.

histBorder

A string with the color of the border of the histogram bars. Applies to plot_ppredictive with features histogram.

histCol

A string with the color of the area of the histogram bars. Applies to plot_ppredictive with features histogram.

histLine

A string with the color of the vertical line related to the fun argument of the plot_ppredictive) function. The line is located at the value returned by fun applied to the actual sample. Applies to plot_ppredictive with features histogram.

ksBorder

A string with the color of the border of the histogram bars. Applies to plot_ppredictive with features ks.

ksCol

A string with the color of the area of the histogram bars. Applies to plot_ppredictive with features ks.

scatterY

A string with the color of the point for the observed sample. Applies to plot_ppredictive with features scatterplot.

scatterYPred

A string with the color of the points for the many samples drawn from the posterior predictive density. Applies to plot_ppredictive with features scatterplot.

seriesY

A string with the color of the observation time series. Applies to plot_series and plot_state_probability.

shadeAlpha

A numeric value between zero and one with the transparency setting for the shaded features, where lower values lead to more transparent shading. Applies to plot_series and plot_state_probability.

states

A vector of strings with the colors of each latent state. Applies to plot_series and plot_state_probability.

Note that the overall legend uses the colors densityY and densityYPred for the actual sample and the posterior predictive samples respectively.

BayesHMM.print: settings for text printouts.

char

A character to use to print text lines and boxes. It defaults to an underscore.

tab

A string to be used for indentation. It defaults to two spaces.

textWidth

An integer with the width of the text lines and boxes. It defaults to 80 characters.

See also